refactor: move cfd components to its own folder
