+++ packages/wallets/src/components/ExternalTradingPlatforms/AddedMT5AccountsList/AddedMT5AccountsList.scss
