+++ packages/wallets/src/components/ExternalTradingPlatforms/AddedMT5AccountsList/AddedMT5AccountsList.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useMT5AccountsList } from '@deriv/api';
-import DerivedMT5 from '../../public/images/mt5-derived.svg';
-import FinancialMT5 from '../../public/images/mt5-financial.svg';
-import SwapFreeMT5 from '../../public/images/mt5-swap-free.svg';
-import { PrimaryActionButton } from '../PrimaryActionButton';
-import { TradingAccountCard } from '../TradingAccountCard';
+import DerivedMT5 from '../../../public/images/mt5-derived.svg';
+import FinancialMT5 from '../../../public/images/mt5-financial.svg';
+import SwapFreeMT5 from '../../../public/images/mt5-swap-free.svg';
+import { PrimaryActionButton } from '../../PrimaryActionButton';
+import { TradingAccountCard } from '../../TradingAccountCard';
 import './AddedMT5AccountsList.scss';
 
 const market_type_to_name_mapper = {
