+++ packages/wallets/src/components/ExternalTradingPlatforms/AvailableDerivezAccountsList/AvailableDerivezAccountsList.scss
