+++ packages/wallets/src/components/ExternalTradingPlatforms/AvailableMT5AccountsList/AvailableMT5AccountsList.scss
