+++ packages/wallets/src/components/ExternalTradingPlatforms/AvailableMT5AccountsList/AvailableMT5AccountsList.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useSortedMT5Accounts } from '@deriv/api';
-import DerivedMT5 from '../../public/images/mt5-derived.svg';
-import FinancialMT5 from '../../public/images/mt5-financial.svg';
-import SwapFreeMT5 from '../../public/images/mt5-swap-free.svg';
-import { SecondaryActionButton } from '../SecondaryActionButton';
-import { TradingAccountCard } from '../TradingAccountCard';
+import DerivedMT5 from '../../../public/images/mt5-derived.svg';
+import FinancialMT5 from '../../../public/images/mt5-financial.svg';
+import SwapFreeMT5 from '../../../public/images/mt5-swap-free.svg';
+import { SecondaryActionButton } from '../../SecondaryActionButton';
+import { TradingAccountCard } from '../../TradingAccountCard';
 import './AvailableMT5AccountsList.scss';
 
 const market_type_to_description_mapper = {
