+++ packages/wallets/src/components/ExternalTradingPlatforms/CFDList/CFDList.tsx
