+++ packages/wallets/src/components/ExternalTradingPlatforms/CFDList/index.ts
