+++ packages/wallets/src/components/ExternalTradingPlatforms/CTraderList/CTraderList.scss
