+++ packages/wallets/src/components/ExternalTradingPlatforms/CTraderList/index.ts
