+++ packages/wallets/src/components/ExternalTradingPlatforms/MT5List/MT5List.scss
