+++ packages/wallets/src/components/ExternalTradingPlatforms/OtherCFDPlatformsList/OtherCFDPlatformsList.scss
