+++ packages/wallets/src/components/ExternalTradingPlatforms/OtherCFDPlatformsList/OtherCFDPlatformsList.tsx
