+++ packages/wallets/src/components/ExternalTradingPlatforms/index.ts
@@ -1,0 +1,9 @@
+export * from './AddedDxtradeAccountsList';
+export * from './AddedMT5AccountsList';
+export * from './AvailableDerivezAccountsList';
+export * from './AvailableDxtradeAccountsList';
+export * from './AvailableMT5AccountsList';
+export * from './CFDList';
+export * from './CTraderList';
+export * from './MT5List';
+export * from './OtherCFDPlatformsList';
