+++ packages/wallets/src/components/index.ts
@@ -1,16 +1,8 @@
 export * from './AccountsList';
-export * from './AddedDxtradeAccountsList';
-export * from './AddedMT5AccountsList';
-export * from './AvailableDerivezAccountsList';
-export * from './AvailableDxtradeAccountsList';
-export * from './AvailableMT5AccountsList';
-export * from './CFDList';
-export * from './CTraderList';
 export * from './DesktopWalletsList';
+export * from './ExternalTradingPlatforms';
 export * from './Loader';
-export * from './MT5List';
 export * from './OptionsAndMultipliersListing';
-export * from './OtherCFDPlatformsList';
 export * from './PrimaryActionButton';
 export * from './ProgressBar';
 export * from './SecondaryActionButton';
